update decimal to binary conversion prototype

+++ holberton.h
@@ -6,6 +6,6 @@ int _putchar(char c);
 int _print_str(char *string);
 int _print_int(long int var);
 int _printf(const char *format, ...);
-long to_Binary(int n);
+int to_Binary(unsigned int n);
 
 #endif
